requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
openai==1.30.0
python-dotenv==1.0.1
lxml==5.2.2
pytest==8.2.0
pytest-asyncio==0.23.7
pgvector==0.2.5
